# backend/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Houji Phase 1 API
  version: 1.0.0-draft
  description: |
    Frozen V1 API draft.
    Backend is the single source of truth for runtime, billing, concurrency, and safety decisions.
    Frontends must not derive rules locally.
    AI endpoints are limited to query, handoff, and work-order submission only.
servers:
  - url: http://127.0.0.1:3000/api/v1
paths:
...
  /device-types:
    get: { summary: Device-type list, tags: [device-type], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListEnvelope' } } } } } }
    post: { summary: Create device type, tags: [device-type], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceTypeUpsert' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericEnvelope' } } } } } }
  /device-types/{id}: { patch: { summary: Update device type, tags: [device-type], parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceTypeUpsert' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericEnvelope' } } } } } } }
...
  /devices/{id}:
    get: { summary: Device detail, tags: [device-ledger], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericEnvelope' } } } } } }
    patch: { summary: Update device, tags: [device-ledger], parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceUpsert' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericEnvelope' } } } } } }
...
components:
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
  schemas:
    EnvelopeBase:
      type: object
      required: [requestId, code, message, data]
      properties:
        requestId: { type: string }
        code: { type: string }
        message: { type: string }
    GenericEnvelope:
      allOf:
        - { $ref: '#/components/schemas/EnvelopeBase' }
        - type: object
          properties:
            data: { type: object, additionalProperties: true }
    ListEnvelope:
      allOf:
        - { $ref: '#/components/schemas/EnvelopeBase' }
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: { type: object, additionalProperties: true }
...
    DeviceTypeUpsert: { type: object, properties: { typeCode: { type: string }, typeName: { type: string }, family: { type: string }, capabilityJson: { type: object, additionalProperties: true }, defaultConfigJson: { type: object, additionalProperties: true } } }
    DeviceUpsert: { type: object, properties: { deviceTypeId: { type: string }, regionId: { type: string }, deviceCode: { type: string }, deviceName: { type: string }, serialNo: { type: string }, protocolType: { type: string } } }
...